requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.0
aiokafka==0.10.0
tenacity==9.0.0
python-dotenv==1.0.1
orjson==3.10.7
uvloop==0.20.0
neo4j==5.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
pgvector==0.2.5
numpy==2.1.1
pandas==2.2.3
scikit-learn==1.5.2
openai==1.47.0
python-jose==3.3.0
cryptography==43.0.1
websockets==13.0
opentelemetry-sdk==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-instrumentation-httpx==0.47b0
loguru==0.7.2
structlog==24.4.0